update image list

+++ views/ck-image/_imageList.php
@@ -2,7 +2,6 @@
 
 use istvan0304\imagemanager\models\CkImage;
 use yii\helpers\Url;
-use yii\widgets\Pjax;
 use yii\helpers\StringHelper;
 
 /* @var $this \yii\web\View */
@@ -10,21 +9,28 @@ use yii\helpers\StringHelper;
 
 ?>
 
-<?php Pjax::begin(['id' => 'ck-pjax-image-list', 'options' => ['class' => 'ck-image-list']]); ?>
-<?php foreach ($ckImages as $ckImage): ?>
-    <?php if ($ckImage->isExistsFile()): ?>
-
-        <div class="ck-image-container">
-            <div class="ck-img-box" data-id="<?= $ckImage->id ?>">
-                <img src="<?= Url::to(['ck-image/preview-thumbnail', 'id' => $ckImage->id]) ?>" class="ck-img" alt="">
-                <p class="ck-image-name"
-                   title="<?= $ckImage->orig_name ?>"><?= StringHelper::truncate($ckImage->orig_name, 15) ?><br>
-                    <span class="ck-img-data"><?= $ckImage->cr_date ?></span><br>
-                    <span class="ck-img-data"><?= CkImage::formatSizeUnits($ckImage->size) ?></span>
-                </p>
+<?php if (!empty($ckImages)): ?>
+    <?php foreach ($ckImages as $ckImage): ?>
+        <?php if ($ckImage->isExistsFile()): ?>
+            <div class="ck-image-container">
+                <div class="ck-img-box" data-id="<?= $ckImage->id ?>">
+                    <img src="<?= Url::to(['ck-image/preview-thumbnail', 'id' => $ckImage->id]) ?>" class="ck-img"
+                         alt="">
+                    <p class="ck-image-name"
+                       title="<?= $ckImage->orig_name ?>"><?= StringHelper::truncate($ckImage->orig_name, 15) ?><br>
+                        <span class="ck-img-data"><?= $ckImage->cr_date ?></span><br>
+                        <span class="ck-img-data"><?= CkImage::formatSizeUnits($ckImage->size) ?></span>
+                    </p>
+                </div>
             </div>
-        </div>
+        <?php endif; ?>
+    <?php endforeach; ?>
+<?php else: ?>
+    <div id="no-image">
+        <div class="no-image-content">
 
-    <?php endif; ?>
-<?php endforeach; ?>
-<?php Pjax::end(); ?>
+            <i class="fas fa-folder-open"></i>
+            <h3><?= Yii::t('ckimage', 'Nothing found.') ?></h3>
+        </div>
+    </div>
+<?php endif; ?>
